Proof of Concept Shell
The commit integrates the kernel in preparation for a shell running on the scheduler. The command line interface first reads each line, second parses and tokenizes it, and finally executes it.

This commit modifies the boot sequence by moving some functionality into the user space. It allows the boot sequence to access BERODS's system call and prevents the kernel from exposing its internals. Furthermore, this update adds the os_open() system call to check if a file exists.